Module Forward_torch.Prms

Deals with arbitrary (nested) parameter records.

Main modules

type ('a, 'const, 'bound) tag =

| Const of 'const

| Free of 'a

| Bounded of 'a * 'bound Base.option * 'bound Base.option

type tagged = (Torch.Tensor.t, Torch.Tensor.t, Torch.Tensor.t) tag

Parameter type, can be constant, free or bounded.

type creator = Torch.Tensor.t -> tagged

Creator creates a parameter type from the original tensor.

type path = Base.String.t Base.List.t

Path specified as a list of strings.

module type Basic = sig ... end

If you want to make your own custom parameter structure which for a reason or another cannot be automated using ppx-prms (e.g. if it is not a record type), the you will need to use Forward_torch.Prms.Make and provide a module of this type.

module type Ops = sig ... end

module type T = sig ... end

Main type for parameter structures.

val value : ('a, 'a, _) tag -> 'a

Extract value from the params t type.

Basic functions to deal with parameter values

val bind : ('a, 'a, 'bound_a) tag -> f:('a -> ('b, 'b, 'bound_b) tag) -> ('b, 'b, 'bound_b) tag

Apply function f to the first component in tagged.

val map : tagged -> f:(Torch.Tensor.t -> Torch.Tensor.t) -> tagged

Apply function f to the first component in tagged and enforce constraints in tagged.

val const : creator

Constructs a constant parameter.

val free : creator

Constructs a free parameter.

val create : ?above:Torch.Tensor.t -> ?below:Torch.Tensor.t -> creator

Constructs a free or (partially) bounded parameter with above and below.

val pin : ?to_:Torch.Tensor.t -> tagged -> tagged

Pin (i.e. convert to Const) either to its current value, or to some to_.

val bound : ?above:Torch.Tensor.t -> ?below:Torch.Tensor.t -> tagged -> tagged

Introduce bounds to a Free parameter with above and below; fails otherwise.

val numel : Torch.Tensor.t -> int

Number of float elements in this parameter.

module Let_syntax : sig ... end

Parameter sets

val cat : string -> path option -> path option

module Make (B : Basic) : T with type 'a p = 'a B.p

This functor takes a Basic module that supplies map, fold, etc, and populates it into a full-fledged T module. If you use the [@@deriving prms] attribute, you won't be needing this.

module None : T with type 'a p = unit

Empty parameter set; useful if your model must expose a parameter module but the model doesn't actually introduce any parameter.

module P : T with type 'a p = 'a

Single parameter

module Pair (P1 : T) (P2 : T) : T with type 'a p = 'a P1.p * 'a P2.p

Pair of parameters

module Option (P : T) : T with type 'a p = 'a P.p Stdlib.Option.t

Single optional parameter

module List (P : T) : T with type 'a p = 'a P.p Stdlib.List.t

List of parameters of the same type.

module Array (P : T) : T with type 'a p = 'a P.p array

Array of parameters of the same type.