Module Prms.Let_syntax

val (let*) : ('a, 'a, _) tag -> ('a -> ('b, 'b, 'bound) tag) -> ('b, 'b, 'bound) tag

Extract value from x and apply function on x (returns unwrapped).

val (and*) : ('a, 'a, 'bound_a) tag -> ('b, 'b, _) tag -> ('a * 'b, 'a * 'b, 'bound_a) tag

Pair each element in x and y and return as Const.

val (let+) : ('a, Torch.Tensor.t, Torch.Tensor.t) tag -> ('a -> Torch.Tensor.t) -> (Torch.Tensor.t, Torch.Tensor.t, Torch.Tensor.t) tag

Map function over x which is wrapped and (returns wrapped).

val (and+) : ('a, Torch.Tensor.t, Torch.Tensor.t) tag -> (Torch.Tensor.t, Torch.Tensor.t, Torch.Tensor.t) tag -> ('a * Torch.Tensor.t, Torch.Tensor.t, Torch.Tensor.t) tag

Pair each element in x and y using only the constraint on x but inject value from y.

val (and++) : ('a, Torch.Tensor.t, Torch.Tensor.t) tag -> Torch.Tensor.t -> ('a * Torch.Tensor.t, Torch.Tensor.t, Torch.Tensor.t) tag

Pair each element in x with y. similar to ( and+ ) but y is Tensor.t instead of tagged.