Module P.M

Module T contains parameters with elements of type Maths.t.

include Ops with type t = Maths.t p and type elt = Maths.t

type elt = Maths.t

type t = Maths.t p

val zeros_like : t -> t

zeros_of x has the same structure as x but each AD.t parameter is replaced by zeros of the same shape.

val ones_like : t -> t

ones_of x has the same structure as x but each AD.t parameter is replaced by ones of the same shape.

val gaussian_like : ?mu:Base.float -> ?sigma:Base.float -> t -> t

gaussian_of ~mu ~sigma x has the same structure as x but each AD.t parameter is replaced by independent normal samples of the same shape (mean ?(mu=0.) and stdev. ?(sigma=1.)).

val gaussian_like_k : ?mu:Base.float -> ?sigma:Base.float -> k:Base.int -> t -> t

Similar to gaussian_like, except for that the tensors sampled has an extra dimension k in front.

val numel : t -> Base.int

Returns the number of parameters in x.

val dot_prod : t -> t -> elt

Returns the dot product between x and y.

val sqr : t -> t

Each element in x is squared.

val sqrt : t -> t

Each element in x is square-rooted.

val (*) : t -> t -> t

Element-wise multiplication of two parameter sets

val (+) : t -> t -> t

Element-wise addition of two parameter sets

val (-) : t -> t -> t

Element-wise subtraction of two parameter sets

val (/) : t -> t -> t

Element-wise division of two parameter sets

val ($*) : Base.float -> t -> t

Multiplication of a parameter set with a scalar

val ($+) : Base.float -> t -> t

Adds a scalar to a parameter set