Parameter Pair.P1

include Basic

type 'a p

val map : 'a p -> f:('a -> 'b) -> 'b p

Apply function f on all elements in x.

val map2 : 'a p -> 'b p -> f:('a -> 'b -> 'c) -> 'c p

Apply function f on all elements in x and y.

val fold : ?path:path -> 'a p -> init:'c -> f:('c -> ('a * path Base.option) -> 'c) -> 'c

Fold x with init and function f. ?path optionally contains a record of the path (a list of strings, in reverse order) taken to arrive at the current value, if it is nested within a broader structure. Make sure to use this if you want to attach string labels to the various components of your custom 'a p type (e.g. use for saving to files, see Forward_torch.Prms.T.T.save_npy).

val fold2 : ?path:path -> 'a p -> 'b p -> init:'c -> f:('c -> ('a * 'b * path Base.option) -> 'c) -> 'c

Fold x and y with init and function f.

type nonrec tagged = tagged p

val iter : 'a p -> f:('a -> Base.unit) -> Base.unit

Apply function f to each element in x.

val iter2 : 'a p -> 'b p -> f:('a -> 'b -> Base.unit) -> Base.unit

Apply function f to each element in x and y.

module T : sig ... end

Module T contains parameters with elements of type Tensor.t.

module M : sig ... end

Module T contains parameters with elements of type Maths.t.

val const : T.t -> M.t

Apply Maths.t to all elements in x.

val value : tagged -> T.t

Extract value as Tensor.t in all elements in x.

val primal : M.t -> T.t

Extract primal of all elements in x.

val tangent : M.t -> T.t

Extract tangents (as Tensor.t) of all elemnts in x.

val make_dual : tagged -> t:Maths.tangent p -> M.t

Make dual number of (primal, t as tangent).

module Let_syntax : sig ... end