Module type Prms.Basic

If you want to make your own custom parameter structure which for a reason or another cannot be automated using ppx-prms (e.g. if it is not a record type), the you will need to use Forward_torch.Prms.Make and provide a module of this type.

type 'a p

val map : 'a p -> f:('a -> 'b) -> 'b p

Apply function f on all elements in x.

val map2 : 'a p -> 'b p -> f:('a -> 'b -> 'c) -> 'c p

Apply function f on all elements in x and y.

val fold : ?path:path -> 'a p -> init:'c -> f:('c -> ('a * path Base.option) -> 'c) -> 'c

Fold x with init and function f. ?path optionally contains a record of the path (a list of strings, in reverse order) taken to arrive at the current value, if it is nested within a broader structure. Make sure to use this if you want to attach string labels to the various components of your custom 'a p type (e.g. use for saving to files, see Forward_torch.Prms.T.T.save_npy).

val fold2 : ?path:path -> 'a p -> 'b p -> init:'c -> f:('c -> ('a * 'b * path Base.option) -> 'c) -> 'c

Fold x and y with init and function f.